Move method to API and improve the return to client
